container/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
umap-learn==0.5.7
scikit-learn==1.5.2
numpy==1.26.4
